Music.Score.Import.Sibelius

Documentation

type IsSibelius a = (HasPitches' a, IsPitch a, HasPart' a, Part a ~ Part, HasArticulation' a, Articulation a ~ Articulation, HasDynamic' a, Dynamic a ~ Dynamics, HasHarmonic a, HasText a, HasTremolo a, Tiable a)

This constraint includes all note types that can be constructed from a Sibelius representation.

fromSibelius :: IsSibelius a => SibeliusScore -> Score a

Convert a score from a Sibelius representation.

readSibelius :: IsSibelius a => FilePath -> IO (Score a)

Read a Sibelius score from a file. Fails if the file could not be read or if a parsing error occurs.

readSibeliusMaybe :: IsSibelius a => FilePath -> IO (Maybe (Score a))

Read a Sibelius score from a file. Fails if the file could not be read, and returns Nothing if a parsing error occurs.

readSibeliusEither :: IsSibelius a => FilePath -> IO (Either String (Score a))

Read a Sibelius score from a file. Fails if the file could not be read, and returns Left m if a parsing error occurs.